Added express server and fake sports betting data
Made a express server and created a route to the fake sports betting data generator. Used Faker package to generate fake sports betting data. Faker is set up to generate over unders and spreads based on the sport. Uploaded NFL and NBA Teams. The team includes the the team name, the city and the initials.